fix: 修正 lint 警告与冗余代码, 颜色调整与收尾清理
- 修正 lint 警告与冗余代码, 音频/视频页快捷键收拢共享层
- 颜色调整
- 代码清理

+++ shared/src/sharedUiMain/kotlin/io/legado/app/ui/book/read/page/ReadViewComposable.kt
@@ -1023,7 +1023,7 @@ private fun AutoPageRevealOverlay(
                 modifier = Modifier
                     .fillMaxSize()
                     .drawWithContent {
-                        clipRect(bottom = revealHeight.coerceIn(0f, size.height.toFloat())) {
+                        clipRect(bottom = revealHeight.coerceIn(0f, size.height)) {
                             this@drawWithContent.drawContent()
                         }
                     },
